refac: moved localAnalytics to a diff file
