+++ src/app/demo/LocalAnalytics.tsx
@@ -1,0 +1,11 @@
+// Define Analytics utilities for tracking
+export const LocalAnalytics = {
+    trackPageView: (pageName: string) => {
+      // In a real implementation, this would integrate with analytics provider
+      console.log(`Analytics: Page view - ${pageName}`);
+    },
+    trackEvent: (eventName: string, properties?: Record<string, any>) => {
+      // In a real implementation, this would integrate with analytics provider
+      console.log(`Analytics: Event - ${eventName}`, properties);
+    },
+  };
